+++ package.json
@@ -3,17 +3,6 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-regular-svg-icons": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "@reduxjs/toolkit": "^1.8.3",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^14.2.1",
-    "@types/jest": "^28.1.4",
-    "@types/node": "^18.0.3",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
     "axios": "^0.27.2",
     "chalk": "^5.0.1",
     "flickity": "^3.0.0",
@@ -29,9 +18,9 @@
     "prestart": "node ./configs/welcome.mjs",
     "start": "react-scripts start & npm run server",
     "build": "react-scripts build",
-    "test": "react-scripts test --coverage ",
+    "test": "react-scripts test --coverage",
     "eject": "react-scripts eject",
-    "server": "json-server --watch db.json -p 3001"
+    "server": "json-server --watch ./src/db/db.json -p 3001"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -49,8 +38,19 @@
     ]
   },
   "devDependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-regular-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@reduxjs/toolkit": "^1.8.3",
     "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^12.0.0",
+    "@testing-library/user-event": "^14.2.1",
     "@types/flickity": "^2.2.6",
+    "@types/jest": "^28.1.4",
+    "@types/node": "^18.0.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@types/react-router-dom": "^5.3.3",
     "eslint": "^8.19.0",
     "eslint-plugin-react": "^7.24.0",
