+++ src/Components/sideMenu/sideMenu.module.scss
@@ -1,52 +1,56 @@
 .sidebar {
-  background: linear-gradient(to bottom, #056cdb, #00b3db);
-  height: 100vh;
-  left: 0;
-  position: fixed;
-  top: 0;
-  width: 3.4em;
-  box-shadow: 0 0 5px 2px rgba(0, 0, 0, .6);
-  .logo {
-    color: var(--white, #ffffff);
-    margin: 1em 0 3em;
-    text-indent: -9999px;
+    background: linear-gradient(to bottom, #056cdb, #00b3db);
+    box-shadow: 0 0 5px 2px rgba(0, 0, 0, .6);
+    height: 100vh;
+    left: 0;
+    position: fixed;
+    top: 0;
+    width: 3.4em;
+
+    .logo {
+        color: var(--white, #ffffff);
+        margin: 1em 0 3em;
+        text-indent: -9999px;
+
     svg {
-      width: 70%;
-      height: 100%;
+        height: 100%;
+        width: 70%;
     }
-  }
-  a {
+}
+
+a {
     display: block;
     padding: 0.7em;
+
     .home {
-      padding: 0.9em;
+        padding: 0.9em;
     }
+
     .sock {
-      padding: 0.8em;
+        padding: 0.8em;
     }
-  }
-  li {
+}
+
+li {
     display: flex;
     justify-content: center;
+
     a {
-      display: flex;
-      justify-content: center;
-      margin-bottom: 1em;
-      padding: 0.5em;
-      color: var(--white, #ffffff);
-      opacity: 0.7;
-      transition: all 0.2s;
-      &:hover,
-      &:focus {
-        background: var(--white, #ffffff);
-        border-radius: 3px;
-        box-shadow: 0px 3px 6px var(--gray);
-        color: var(--darkpurple, #343175);
-        opacity: 1;
+        color: var(--white, #ffffff);
+        display: flex;
+        justify-content: center;
+        margin-bottom: 1em;
+        opacity: 0.7;
+        padding: 0.5em;
+        transition: all 0.2s;
+        &:hover,
+        &:focus {
+            background: var(--white, #ffffff);
+            border-radius: 3px;
+            box-shadow: 0px 3px 6px var(--gray);
+            color: var(--darkpurple, #343175);
+            opacity: 1;
       }
     }
-  }
-  img {
-    width: 100%;
   }
 }
